update box and scatter plot
